+++ lib/nightwing/instrumentation/active_record.rb
@@ -1,10 +1,8 @@
-require "nightwing/instrumentation/base"
-
 module Nightwing
   module Instrumentation
     # This code was influenced by Matt Aimonetti's awesome blog post: "Practical Guide to StatsD/Graphite Monitoring"
     # Check it out: https://matt.aimonetti.net/posts/2013/06/26/practical-guide-to-graphite-monitoring
-    class ActiveRecord < Base
+    class ActiveRecord
       SQL_INSERT_DELETE_PARSER_REGEXP = /^(\w+)\s(\w+)\s\W*(\w+)/
       SQL_SELECT_REGEXP = /select .*? FROM \W*(\w+)/i
       SQL_UPDATE_REGEXP = /update \W*(\w+)/i
@@ -15,8 +13,15 @@ module Nightwing
         end
 
         if table
+          action.downcase!
+          table.downcase!
           query_time = (finished - started) * 1_000
-          client.timing("#{namespace}.sql.#{table.downcase}.#{action.downcase}.time", query_time.round)
+
+          Nightwing.client.timing "sql.#{action}.time", query_time.round
+          Nightwing.client.timing "sql.#{action}.#{table}.time", query_time.round
+
+          Nightwing.client.increment "sql.#{action}.processed"
+          Nightwing.client.increment "sql.#{action}.#{table}.processed"
         end
       end