+++ lib/nightwing/sidekiq/base.rb
@@ -1,11 +1,28 @@
 require "sidekiq/api"
-require "nightwing/base"
+require "nightwing/metric"
+require "nightwing/logger"
+require "nightwing/client_logger"
 
 module Nightwing
   module Sidekiq
-    class Base < Nightwing::Base
+    class Base
+      attr_reader :namespace, :logger
+
       def initialize(options = {})
-        super options.merge(namespace: "sidekiq")
+        @namespace = options.fetch(:namespace, "sidekiq")
+        @client = options.fetch(:client, Nightwing.client)
+        @logger = options.fetch(:logger, Nightwing::Logger.new)
+        @debug = options.fetch(:debug, false)
+      end
+
+      def client
+        @client_proxy ||= begin
+          if @debug
+            Nightwing::ClientLogger.new(client: @client, logger: @logger)
+          else
+            @client
+          end
+        end
       end
 
       private