+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+## Unreleased
+
+- New profiling extension for Redis
+- New profiling extension for Memcache (via Dalli)
+- Added new `Nightwing.client` to global statds client configuration
+- Internal setup refactor
+- Metric changes are listed below
+
+**Renamed Metrics**
+
+- `instrumentation.sql.<table>.<action>.time` to `sql.<table>.<action>.time`
+
+**New Metrics**
+
+- `memcache.command.time`
+- `memcache.command.<command>.time`
+- `memcache.command.processed`
+- `memcache.command.<command>.processed`
+- `redis.command.time`
+- `redis.command.<command>.time`
+- `redis.command.processed`
+- `redis.command.<command>.processed`
+
 ## 0.2.0
 
 - Renamed memory_used metric so that we can utilize the memory namespace for additional metrics